requirements.txt
telethon
aiogram==3.*
apscheduler
python-dotenv
